Remove OS specific implementations
