+++ filenames.gypi
@@ -130,9 +130,6 @@
       'atom/browser/api/atom_api_net.h',
       'atom/browser/api/atom_api_notification.cc',
       'atom/browser/api/atom_api_notification.h',
-      'atom/browser/api/atom_api_notification_mac.mm',
-      'atom/browser/api/atom_api_notification_win.cc',
-      'atom/browser/api/atom_api_notification_linux.cc',
       'atom/browser/api/atom_api_power_monitor.cc',
       'atom/browser/api/atom_api_power_monitor.h',
       'atom/browser/api/atom_api_power_save_blocker.cc',
@@ -361,10 +358,6 @@
       'atom/browser/ui/win/notify_icon.h',
       'atom/browser/ui/win/taskbar_host.cc',
       'atom/browser/ui/win/taskbar_host.h',
-      'atom/browser/ui/win/toast_handler.cc',
-      'atom/browser/ui/win/toast_handler.h',
-      'atom/browser/ui/win/toast_lib.cc',
-      'atom/browser/ui/win/toast_lib.h',
       'atom/browser/ui/x/event_disabler.cc',
       'atom/browser/ui/x/event_disabler.h',
       'atom/browser/ui/x/window_state_watcher.cc',
